pyproject.toml
[tool.poetry]
name = "ksenyaChatBotAdmin"
version = "0.1.0"
description = ""
authors = ["Dair Sheishenaliev <dokadair@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^5.0.0"
gunicorn = "^20.1.0"
mysql-connector-python = "^8.0.33"
wheel = "^0.40.0"
django-admin-rangefilter = "^0.10.0"
requests = "^2.30.0"
ipython = "^8.13.2"
django-cors-headers = "^4.0.0"
python-decouple = "^3.8"
pillow = "^9.5.0"
psycopg2 = "^2.9.6"
whitenoise = "^6.5.0"
celery = "^5.3.4"
redis = "^5.0.1"
djangorestframework = "^3.15.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
